+++ Language/Haskell/TH/Desugar/Reify.hs
@@ -85,7 +85,7 @@ getDataD err name = do
            TyConI dec -> return dec
            _ -> badDeclaration
   case dec of
-#if MIN_VERSION_template_haskell(2,11,0)
+#if __GLASGOW_HASKELL__ > 710
     DataD _cxt _name tvbs _mk cons _derivings -> return (tvbs, cons)
     NewtypeD _cxt _name tvbs _mk con _derivings -> return (tvbs, [con])
 #else
@@ -105,7 +105,7 @@ dataConNameToDataName con_name = do
   case info of
 #if __GLASGOW_HASKELL__ > 710
     DataConI _name _type parent_name -> return parent_name
-#else    
+#else
     DataConI _name _type parent_name _fixity -> return parent_name
 #endif
     _ -> fail $ "The name " ++ show con_name ++ " does not appear to be " ++
@@ -128,7 +128,7 @@ dataConNameToCon con_name = do
     get_con_name (RecC name _)        = [name]
     get_con_name (InfixC _ name _)    = [name]
     get_con_name (ForallC _ _ con)    = get_con_name con
-#if MIN_VERSION_template_haskell(2,11,0)
+#if __GLASGOW_HASKELL__ > 710
     get_con_name (GadtC names _ _)    = names
     get_con_name (RecGadtC names _ _) = names
 #endif
@@ -187,7 +187,7 @@ reifyInDec :: Name -> [Dec] -> Dec -> Maybe Info
 reifyInDec n decs (FunD n' _) | n `nameMatches` n' = Just $ mkVarI n decs
 reifyInDec n decs (ValD pat _ _)
   | any (nameMatches n) (S.elems (extractBoundNamesPat pat)) = Just $ mkVarI n decs
-#if MIN_VERSION_template_haskell(2,11,0)
+#if __GLASGOW_HASKELL__ > 710
 reifyInDec n _    dec@(DataD    _ n' _ _ _ _) | n `nameMatches` n' = Just $ TyConI dec
 reifyInDec n _    dec@(NewtypeD _ n' _ _ _ _) | n `nameMatches` n' = Just $ TyConI dec
 #else
@@ -217,7 +217,7 @@ reifyInDec n _    dec@(ClosedTypeFamilyD n' _ _ _) | n `nameMatches` n'
 #endif
 #endif
 
-#if MIN_VERSION_template_haskell(2,11,0)
+#if __GLASGOW_HASKELL__ > 710
 reifyInDec n decs (DataD _ ty_name tvbs _mk cons _)
   | Just info <- maybeReifyCon n decs ty_name (map tvbToType tvbs) cons
   = Just info
@@ -252,7 +252,7 @@ reifyInDec n decs (InstanceD _ _ sub_decs)
     reify_in_instance dec@(DataInstD {})    = reifyInDec n (sub_decs ++ decs) dec
     reify_in_instance dec@(NewtypeInstD {}) = reifyInDec n (sub_decs ++ decs) dec
     reify_in_instance _                     = Nothing
-#if MIN_VERSION_template_haskell(2,11,0)
+#if __GLASGOW_HASKELL__ > 710
 reifyInDec n decs (DataInstD _ ty_name tys _ cons _)
   | Just info <- maybeReifyCon n decs ty_name tys cons
   = Just info
@@ -296,7 +296,7 @@ maybeReifyCon n decs ty_name ty_args cons
     con_to_type (RecC _ vstys)   = mkArrows (map thdOf3 vstys) result_ty
     con_to_type (InfixC t1 _ t2) = mkArrows (map snd [t1, t2]) result_ty
     con_to_type (ForallC bndrs cxt c) = ForallT bndrs cxt (con_to_type c)
-#if MIN_VERSION_template_haskell(2,11,0)
+#if __GLASGOW_HASKELL__ > 710
     con_to_type (GadtC _ stys rty)     = mkArrows (map snd    stys)  rty
     con_to_type (RecGadtC _ vstys rty) = mkArrows (map thdOf3 vstys) rty
 #endif
@@ -338,7 +338,7 @@ findInstances n = map stripInstanceDec . concatMap match_instance
   where
     match_instance d@(InstanceD _ ty _)        | ConT n' <- ty_head ty
                                                , n `nameMatches` n' = [d]
-#if MIN_VERSION_template_haskell(2,11,0)
+#if __GLASGOW_HASKELL__ > 710
     match_instance d@(DataInstD _ n' _ _ _ _)    | n `nameMatches` n' = [d]
     match_instance d@(NewtypeInstD _ n' _ _ _ _) | n `nameMatches` n' = [d]
 #else
@@ -401,7 +401,7 @@ findCon n = find match_con
     match_con (RecC n' _)     = n `nameMatches` n'
     match_con (InfixC _ n' _) = n `nameMatches` n'
     match_con (ForallC _ _ c) = match_con c
-#if MIN_VERSION_template_haskell(2,11,0)
+#if __GLASGOW_HASKELL__ > 710
     match_con (GadtC nms _ _)    = any (n `nameMatches`) nms
     match_con (RecGadtC nms _ _) = any (n `nameMatches`) nms
 #endif